Fixed bug with weapon components. Partial EntitiesPed.cs
