+++ Client/CoopClient.csproj
@@ -80,6 +80,8 @@
     <Compile Include="Entities\EntitiesPed.cs" />
     <Compile Include="Entities\EntitiesPlayer.cs" />
     <Compile Include="Entities\EntitiesThread.cs" />
+    <Compile Include="Entities\Sync\OnFootSync.cs" />
+    <Compile Include="Entities\Sync\VehicleSync.cs" />
     <Compile Include="Main.cs" />
     <Compile Include="Menus\MenusMain.cs" />
     <Compile Include="Menus\Sub\Servers.cs" />
