+++ Client/Packets.cs
@@ -410,10 +410,10 @@ namespace CoopClient
         public LVector3 AimCoords { get; set; }
 
         [ProtoMember(8)]
-        public int CurrentWeaponHash { get; set; }
+        public uint CurrentWeaponHash { get; set; }
 
         [ProtoMember(9)]
-        public List<uint> WeaponComponents { get; set; }
+        public Dictionary<uint, bool> WeaponComponents { get; set; }
 
         [ProtoMember(10)]
         public byte? Flag { get; set; } = 0;
@@ -564,7 +564,7 @@ namespace CoopClient
         public LVector3 AimCoords { get; set; }
 
         [ProtoMember(6)]
-        public int CurrentWeaponHash { get; set; }
+        public uint CurrentWeaponHash { get; set; }
 
         [ProtoMember(7)]
         public byte? Flag { get; set; } = 0;
@@ -855,7 +855,7 @@ namespace CoopClient
         public LVector3 AimCoords { get; set; }
 
         [ProtoMember(10)]
-        public int CurrentWeaponHash { get; set; }
+        public uint CurrentWeaponHash { get; set; }
 
         [ProtoMember(11)]
         public byte? Flag { get; set; } = 0;
